tests work now WAHOOO
