+++ package.json
@@ -1,28 +1,35 @@
 {
-  "name": "srfline",
-  "version": "1.0.0",
-  "description": "surfline full stack app",
-  "main": "server.js",
-  "homepage": "https://stormy-thicket-96949.herokuapp.com/",
-  "scripts": {
-    "start": "nodemon server.js",
-    "build": "cd client/ && npm install && npm run build"
-  },
-  "author": "Jboe",
-  "license": "MIT",
-  "dependencies": {
-    "@hapi/joi": "^17.1.1",
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.19.0",
-    "concurrently": "^5.3.0",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "heroku": "^7.47.2",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.10.13",
-    "morgan": "^1.10.0",
-    "nodemon": "^2.0.6"
-  },
-  "devDependencies": {}
+	"name": "srfline",
+	"version": "1.0.0",
+	"description": "surfline full stack app",
+	"main": "server.js",
+	"homepage": "https://stormy-thicket-96949.herokuapp.com/",
+	"scripts": {
+		"start": "nodemon server.js",
+		"build": "cd client/ && npm install && npm run build",
+		"backTest": "jest src/app.test.js --watchAll --detectOpenHandles"
+	},
+	"author": "Jboe",
+	"license": "MIT",
+	"dependencies": {
+		"@hapi/joi": "^17.1.1",
+		"bcryptjs": "^2.4.3",
+		"body-parser": "^1.19.0",
+		"concurrently": "^5.3.0",
+		"cors": "^2.8.5",
+		"dotenv": "^8.2.0",
+		"express": "^4.17.1",
+		"heroku": "^7.47.2",
+		"jsonwebtoken": "^8.5.1",
+		"mongoose": "^5.10.13",
+		"morgan": "^1.10.0",
+		"nodemon": "^2.0.6"
+	},
+	"devDependencies": {
+		"jest": "^27.0.6",
+		"supertest": "^6.1.3"
+	},
+	"jest": {
+		"testEnvironment": "node"
+	}
 }
