+++ src/resources/spots/spot.route.js
@@ -5,7 +5,8 @@ const router = express.Router();
 
 router.route('/').get(controllers.getMany).post(controllers.createOne);
 
-router.route('/:id')
+router
+	.route('/:id')
 	.get(controllers.getOne)
 	.put(controllers.updateOne)
 	.delete(controllers.removeOne);
