+++ src/resources/users/user.route.js
@@ -3,7 +3,8 @@ const controllers = require('./user.controller');
 const router = express.Router();
 const verify = require('../verifyToken');
 
-router.use(verify.tokenAuth)
+router
+	.use(verify.tokenAuth)
 	.route('/')
 	.put(controllers.updateOnePrime)
 	.get(controllers.getOneAndPopulateSpot)
