+++ src/utils/time.js
@@ -1,0 +1,3 @@
+exports.setTimeout = function () {
+	return global.setTimeout.apply(global, arguments);
+};
